requirements.txt
django==2.2
psycopg2==2.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
django-cors-headers==3.4.0